Moved functions to App.js
