+++ code/src/components/Thoughts.js
@@ -1,45 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
-import Thought from 'components/thought/Thought';
-import NewInput from 'components/newthought/NewInput';
-import { THOUGHTS_URL } from '../urls';
+import Thought from './thought/Thought';
+import NewInput from './newthought/NewInput';
 
-const Thoughts = () => {
-  const [thoughts, setThoughts] = useState([]);
-
-  useEffect(() => {
-    fetchThoughts();
-  }, []);
-
-  const fetchThoughts = () => {
-    fetch(THOUGHTS_URL)
-      .then(response => response.json())
-      .then(json => {
-        const filteredThoughts = json.filter(thought => thought.message);
-        const limitedThoughts = filteredThoughts.slice(0, 20);
-        setThoughts(limitedThoughts);
-      })
-      .catch(error => console.error(error));
-  };
-
-  const addNewThought = newThought => {
-    fetch(THOUGHTS_URL, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ message: newThought })
-    }).then(() => fetchThoughts());
-  }
-
-  const onLiked = thoughtId => {
-    const updatedThoughts = thoughts.map(thought => {
-      if (thought._id === thoughtId) {
-        thought.hearts += 1
-      }
-      return thought
-    })
-    setThoughts(updatedThoughts)
-  }
-
+const Thoughts = ({ thoughts, onLiked, addNewThought }) => {
   return (
     <section>
       <NewInput onInputChange={addNewThought} />
