+++ code/src/components/newthought/Description.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Description = () => {
+  return (
+    <p className="new-message">What's making you happy right now?</p>
+  )
+}
+
+export default Description;
