+++ code/src/components/thought/Message.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
-const Message = (props) => {
+const Message = ({ message }) => {
   return (
     <p className="message">
-      {props.message}
+      {message}
     </p>
   )
 }
