--- code/src/urls.js
@@ -1,1 +1,0 @@
-export const THOUGHTS_URL = "https://happy-thoughts-technigo.herokuapp.com/thoughts";
